🩻 Create better separation for the models
Previously, the same module was being used across all the application.
Now, a new "core.model" was created to share this module in the
"domain", "repository" and "service" layers. This is a bit different
from what I usually implement, but I'm trying something new here to see
how it goes.