🐳 Update configuration files to fix Docker Image
After implementing the HOCON Config file to save the API keys, the
project ended up with two `application.*` files, one YAML and one CONF.
This created some issues with Docker, that was trying to get the YAML to
use with Hocon. The solution (and best practice) was to have only a
single file for configuration. A new file exclusive for secrets were
created and .gitignore and README.md were updated to reflect this
change.