📝 Update README with local Docker run instructions
Now the Firestore is integrated, additional steps needs to be taken to
run the Docker image locally. The Google Cloud Default Credential is
required to be in the image. When the image is on Google Cloud Run, the
service will get the JSON automatically, so no extra step is needed.