🚚 Move the server code to its dedicated module
Savel will be a monorepo project with Kotlin Multiplatform modules for
different platforms. For this reason, the server-specific code was moved
to a dedicated module and the reference files were updated to reflect
this change, such as the Dockerfile.

The Shadow plugin was added because the `fatJar` and `shadowJar` weren't
working properly after the change to KMP.